FIX: Error construction for APIs
